tf2 between camera and map frames
It prints the coordinates w.r.t. map frame.
We need to create a list of apriltags we are interested in and insert their coordinates when we detect them.
Careful! there is some noise when we transorm the coordinates, Maybe compute the average? Or just insert the first value?

Also we have to code the idea of
- I see red, I go towards it